HW7 (PageObject) corrected. Fixtures moved to conftest.py
